Delete the old client and transaction, refactor and change the name of the new client and transaction
